Move PackageFiltersRepository into its own file.
