+++ core/src/main/scala/org/genivi/sota/core/rvi/SotaServices.scala
@@ -1,7 +1,7 @@
 /**
-  * Copyright: Copyright (C) 2015, Jaguar Land Rover
-  * License: MPL-2.0
-  */
+ * Copyright: Copyright (C) 2015, Jaguar Land Rover
+ * License: MPL-2.0
+ */
 package org.genivi.sota.core.rvi
 
 import akka.actor.{ActorRef, ActorSystem}
